Created Home View and viewController
The home view and its control were created based on the model already provided. The extension UITableViewDelegate was created and UITableViewDataSource in HomeviewController. Also created CodeView protocol with two functions: ** setupViews and setupConstraints **, this protocol must be inserted has all views created programmatically for ease of understanding. Two extensions have also been created for UIColor, one for creating colors that will be used in the app based on ios color mode with navBarColor and titleNavBar Color being created, and another for converting a default color of 0..255 to 0. ..1.