Add testUi in Home and FactTableViewCell
FactViewCell and HomeScreenTest classes have been added to test how these screens will interface if they fit the model. The SizeViewTest enum was also inserted to generate the size of a view to be tested and returns the size of the view. for example testing an interface on an iphone 8 or X. A protocol was created called SizeViewTest, which has the function createViewForTest () whose purpose is to return a view to be tested by Nimble_Snapshots. To be more general, a UIViewForSnapShot class was created that receives the view to be tested and applies the settings for the test, such as portrait mode.