Add uiTest
Quick, Nimble and Nimble_Snapshots Pods have been added to perform snepshot testing to check the integrity of Home view and FactTableViewCell components to check integrity if layout changes occur. An enum has been created to generate the test view and can be placed on different devices based on size. and the UIViewForSnapShot class to simulate the device in view along with a test protocol that requires creating the test view.