Connect functions between view and controller
The view home search delegate (uiTextFiel) has been added which is responsible for issuing when the user clicks return from the keyboard, as well as two notifications in SearchFactViewController to raise the view when the keyboard is called or to drop it when the keyboard leaves, to prevent the keyboard from being on the saw. RXswift was also inserted to connect SearchFactViewController and HomeViewController using the textForSearch variable which is of type PublishSubject. Finally the view search colors were updated in the following components:
viewModal and buttonSearch, using the colors added in Uicolor + AppColor, with the new colors being created borderButtonColor, textButtonColor buttonColor.