Created from network layer
The protocol-focused network layer was created. The ServiceProtocol protocol was created which shows all the components that are required to perform the api request. ProviderProtocol which is the function that performed the resquet. HTTPMethod, HTTPMethod, ParametersEncoding, NetworkResponse, and NetworkError were created only to componentize these values that can be modified as needed.
The URLComponents extension was also created to join baseURL with the path and add encoding and URLRequest for those that don't need to query.
Procoloco URLSessionProtocol and URLSessionDataTaskProtocol are designed to facilitate mockup-based testing.