Perform text search
The NetworkService class was created to enter the initial configuration to perform the api search. As well as creating a getFact function in HomeViewController to perform the search according to the text entered on the SearchFact screen, passed via RXSwift. Two structures were also created to work with json data, the first factDTO that receives json which is an array called result and the FACT structure that receives sim data from fact as value and category.