Add NetWorkLayer Test Layer
First the URLSessionProtocol return type was modified to URLSessionDataTaskProtocol so that you can create a Mock Type file for testing. It was creator folder named NetWorkLayerTest to store Mock`s files. Was raised :
URLSessionDataTaskMock to simulate URLSessionDataTaskProtocol;
ModelMock to simulate the return data type. ServiceMock for the possible request types and their results;
URLSessionMock to simulate the server which returns the values ​​obtained from the internet;
NetworkLayerTest to perform the appropriate tests, with the following categories:
Check integrity of the parameters for making a request;
Check all cases of Request;
Check results obtained after request.
Also, a json file called Facts.json was added, obtained from an api request to check if the return conforms to the structures, thus avoiding a modification in the structure that may give an error in the future in the application.