Create Alert Errors and check internet connection
An enemy TypeErros was created to check the types of errors that may occur in the application, initially there are three:
notNetWork - Concerning Internet Concern
notConnectServe - for server retonation types
textLessThan3 - When user tries to submit a search with less than 3 words, it was created to avoid a request as it will be possible to search locally.

NOTE: For this purpose the search keyboard type has been changed to asciiCapable, preventing users from sending emojis to search.

The AlertsError class was also created which creates an alert and shows the user according to their type.
Finally the Network class was created to check if the device is connected to the internet.