Add error alert
An error alert has been added when the user searches for text shorter than 3 characters in SearchFactViewController as well as validating if it is connected to the internet or not. Also a new type of error was added invalidSearch, for when the user searches with symbols which is not recognized by api. Checked out when launching the app, and only checks when performing a search.