Add character and category check
The FontSizeFact enum has been added which returns the size of the text that can be displayed in fact being 17 for larger than 80 and 24 for smaller.
It was also created the class FactSettings that aims to manipulate objects of type Fact and returns the category in uppercased and if not returns UNCATEGORIZED and finally return the font size of the fact to be presented.