Add protocol datasource and class FactDatasource
A protocol was created to create the FactDatasource class that will be responsible for manipulating the table data so that it is not the sole responsibility of HomeViewController, decoupling this functionality, as well as a setup in FactTableViewCell function that can receive the values shown in it as an injection. dependency, respecting the encapsulation.