Added Facs in Home
facts were added to the homeViewController table via the setupTableView function, as well as added functionality that returns to the top of the table every time a new request is made to the ItemsTableViewDatasource. And change category color to suit dark mode by creating a new color in assst`s categoryColor.