Add sharing fact
A buttonAction (sender: UIButton!) Function has been added in FactDatasource that associates an action with the button that modifies the sharingFact variable which is of type PublishSubject which sends the data to HomeViewController with the right fact, and it calls showSharingFact which calls activitityController which It is native and allows sharing natively between options.
Note: If shared via whatsApp, it generates an error because it is not native, but sends the same way, because it passes this responsibility to whatsapp.