Add initial feedback and notResult
Two enums were created to create the screens. The first is FeedBackProperties which returns a tuple with the text and image that will be inserted into the FeedBack view. the second enum is SetupViewInHome which has the three possible screens that are in the home, the home having a table, the initial being a picture and photo and the not result.
A color has also been created for the text called feedBackColor. Along with their respective SnepShot tests FeedBackINotResultScreen and FeedBackIInitialScreen