Add UIActivityIndicatorView
A UIActivityIndicatorView has been added to give user feedback that data is being loaded. Due to the fact that the main screen views are always changing for better visual feedback to the user, the setupActivityView function was created to set the constraints correctly as well as the initial component settings such as size and positioning.