Rest API Integration Test
The integration test for get API along with local json has been refactored to verify the integrity of the call and the result so that it can be passed to the table. Some Enum tests were also inserted to check the integrity of the images to be placed as well as the text.